AssessmentDAO => AssessmentRepository and associated cleanup

+++ caliber/src/main/java/com/revature/caliber/controllers/AssessmentController.java
@@ -111,7 +111,7 @@ public class AssessmentController {
 	@Transactional(isolation = Isolation.READ_COMMITTED, propagation = Propagation.REQUIRED)
 	@PreAuthorize("hasAnyRole('VP', 'TRAINER', 'STAGING')")
 	public ResponseEntity<List<Assessment>> findAssessmentByWeek(@PathVariable Integer batchId,
-			@PathVariable Integer week) {
+			@PathVariable Short week) {
 		log.debug("Find assessment by week number " + week + " for batch " + batchId + " ");
 		List<Assessment> assessments = assessmentService.findAssessmentByWeek(batchId, week);
 		if(assessments.isEmpty()){
